# api-swagger-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "API Documentation",
        "contact": {
            "email": "viko.ramayudha@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API Documentation"
        }
    ],
    "paths": {
        "/mahasiswa": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mahasiswa - Get All",
                "description": "Mengambil Data Mahasiswa",
                "operationId": "listMahasiswa",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil mengambil Data Mahasiswa",
                                    "data": [
                                        {
                                            "mhs_nim": "80224020",
                                            "mhs_nama": "Mukhamad Viko Ramayudha"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mahasiswa/{id}": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mahasiswa - Get By NIM",
                "description": "Mengambil Data Mahasiswa Berdasarkan NIM",
                "operationId": "listMahasiswaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Masukkan NIM Mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "80224020"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil mengambil Data Mahasiswa",
                                    "data": {
                                        "mhs_nim": "80224020",
                                        "mhs_nama": "Mukhamad Viko Ramayudha"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data Mahasiswa tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Data Mahasiswa tidak ditemukan"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mahasiswa/create": {
            "post": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mahasiswa - Insert",
                "description": "Menambahkan Data Mahasiswa",
                "operationId": "insertMahasiswa",
                "requestBody": {
                    "description": "Data Mahasiswa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "mhs_nim": "80224020",
                                "mhs_nama": "Mukhamad Viko Ramayudha"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil menambahkan Data Mahasiswa",
                                    "data": {
                                        "mhs_nim": "80224020",
                                        "mhs_nama": "Mukhamad Viko Ramayudha"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "mhs_nim": [
                                        "The mhs nim field is required."
                                    ],
                                    "mhs_nama": [
                                        "The mhs nama field is required."
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mahasiswa/update": {
            "put": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mahasiswa Update",
                "description": "Update data Mahasiswa",
                "operationId": "updateMahasiswa",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "mhs_nim": "80224020",
                                "mhs_nama": "Mukhamad Viko Ramayudha"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil diubah"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mahasiswa/delete": {
            "delete": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mahasiswa Delete",
                "description": "Hapus data Mahasiswa",
                "operationId": "deleteMahasiswa",
                "requestBody": {
                    "description": "ID Mahasiswa yang ingin dihapus",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "mhs_nim": "80224020"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil dihapus"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token",
                "scheme": "bearer"
            }
        }
    }
}